refactor(product-card): rename component
