+++ components/index.js
@@ -41,7 +41,7 @@ export { default as PaymentMethodIcon } from './payment-method-icon';
 export { default as PaymentMethodRadioButtonCard } from './payment-method-radio-button-card';
 export { default as PointsRedemption } from './points-redemption';
 export { default as ProductDetails } from './product-details.vue';
-export { default as ProductOrderCard } from './product-order-card.vue';
+export { default as ProductCard } from './product-card.vue';
 export { default as QuantitySelector } from './quantity-selector';
 export { default as QrCode } from './qr-code';
 export { default as RadioButton } from './radio-button';
